requirements.txt
discord.py==2.1.1
paypalrestsdk==1.13.1
aiosqlite==0.18.0
pytz==2022.7.1
expr.py==0.3.0
openai==0.27.0
asgiref==3.6.0
qrcode==7.4.2
flask==2.2.3
pyyaml==6.0